+++ home/models.py
@@ -25,7 +25,6 @@ from wagtail.admin.panels import (
 )
 from wagtail.api import APIField
 from wagtail.blocks import (
-    StreamBlockValidationError,
     StreamValue,
     StructBlockValidationError,
 )
@@ -298,7 +297,6 @@ class WhatsappBlock(blocks.StructBlock):
         "media cannot exceed 1024 characters.",
         validators=(MaxLengthValidator(4096),),
     )
-
     example_values = blocks.ListBlock(
         blocks.CharBlock(
             label="Example Value",
@@ -350,21 +348,8 @@ class WhatsappBlock(blocks.StructBlock):
 
     def clean(self, value: dict[str, Any]) -> dict[str, Any]:
         result = super().clean(value)
-        num_vars_in_msg = len(re.findall(r"{{\d+}}", result["message"]))
         errors = {}
-        example_values = result["example_values"]
-        for ev in example_values:
-            if "," in ev:
-                errors["example_values"] = ValidationError(
-                    "Example values cannot contain commas"
-                )
-        if num_vars_in_msg > 0:
-            num_example_values = len(example_values)
-            if num_vars_in_msg != num_example_values:
-                errors["example_values"] = ValidationError(
-                    f"The number of example values provided ({num_example_values}) "
-                    f"does not match the number of variables used in the template ({num_vars_in_msg})",
-                )
+
         if (result["image"] or result["document"] or result["media"]) and len(
             result["message"]
         ) > self.WHATSAPP_MESSAGE_MAX_LENGTH:
@@ -801,13 +786,6 @@ class ContentPage(UniqueSlugMixin, Page, ContentImportMixin):
     def whatsapp_template_image(self):
         return self.whatsapp_body.raw_data[0]["value"]["image"]
 
-    @property
-    def whatsapp_template_example_values(self):
-        example_values = self.whatsapp_body.raw_data[0]["value"].get(
-            "example_values", []
-        )
-        return [v["value"] for v in example_values]
-
     def create_whatsapp_template_name(self) -> str:
         return f"{self.whatsapp_template_prefix}_{self.get_latest_revision().pk}"
 
@@ -916,7 +894,6 @@ class ContentPage(UniqueSlugMixin, Page, ContentImportMixin):
             self.locale,
             self.whatsapp_template_buttons,
             self.whatsapp_template_image,
-            self.whatsapp_template_example_values,
         )
 
         return self.whatsapp_template_name
@@ -1081,72 +1058,6 @@ class ContentPage(UniqueSlugMixin, Page, ContentImportMixin):
                     ).strip()
 
                     block.value["message"] = cleaned_message
-
-        # The WA title is needed for all templates to generate a name for the template
-        if self.is_whatsapp_template and not self.whatsapp_title:
-            errors.setdefault("whatsapp_title", []).append(
-                ValidationError("All WhatsApp templates need a title.")
-            )
-        # The variable check is only for templates
-        if self.is_whatsapp_template and len(self.whatsapp_body.raw_data) > 0:
-            whatsapp_message = self.whatsapp_body.raw_data[0]["value"]["message"]
-
-            right_mismatch = re.findall(r"(?<!\{){[^{}]*}\}", whatsapp_message)
-            left_mismatch = re.findall(r"\{{[^{}]*}(?!\})", whatsapp_message)
-            mismatches = right_mismatch + left_mismatch
-
-            if mismatches:
-                errors.setdefault("whatsapp_body", []).append(
-                    StreamBlockValidationError(
-                        {
-                            0: StreamBlockValidationError(
-                                {
-                                    "message": ValidationError(
-                                        f"Please provide variables with matching braces. You provided {mismatches}."
-                                    )
-                                }
-                            )
-                        }
-                    )
-                )
-
-            vars_in_msg = re.findall(r"{{(.*?)}}", whatsapp_message)
-            non_digit_variables = [var for var in vars_in_msg if not var.isdecimal()]
-
-            if non_digit_variables:
-                errors.setdefault("whatsapp_body", []).append(
-                    StreamBlockValidationError(
-                        {
-                            0: StreamBlockValidationError(
-                                {
-                                    "message": ValidationError(
-                                        f"Please provide numeric variables only. You provided {non_digit_variables}."
-                                    )
-                                }
-                            )
-                        }
-                    )
-                )
-
-            # Check variable order
-            actual_digit_variables = [var for var in vars_in_msg if var.isdecimal()]
-            expected_variables = [
-                str(j + 1) for j in range(len(actual_digit_variables))
-            ]
-            if actual_digit_variables != expected_variables:
-                errors.setdefault("whatsapp_body", []).append(
-                    StreamBlockValidationError(
-                        {
-                            0: StreamBlockValidationError(
-                                {
-                                    "message": ValidationError(
-                                        f'Variables must be sequential, starting with "{{1}}". Your first variable was "{actual_digit_variables}"'
-                                    )
-                                }
-                            )
-                        }
-                    )
-                )
 
         if errors:
             raise ValidationError(errors)
@@ -1674,7 +1585,7 @@ class WhatsAppTemplate(
         SUBMITTED = "SUBMITTED", _("Submitted")
         FAILED = "FAILED", _("Failed")
 
-    def get_submission_status_display(self):
+    def get_submission_status_display(self) -> str:
         return self.SubmissionStatus(self.submission_status).label
 
     get_submission_status_display.admin_order_field = "submission status"
@@ -1687,7 +1598,7 @@ class WhatsAppTemplate(
         default=Category.MARKETING,
     )
 
-    def get_category_display(self):
+    def get_category_display(self) -> str:
         return self.Category(self.category).label
 
     get_category_display.admin_order_field = "category"
@@ -1728,7 +1639,6 @@ class WhatsAppTemplate(
     submission_status = models.CharField(
         max_length=30,
         choices=SubmissionStatus.choices,
-        blank=True,
         default=SubmissionStatus.NOT_SUBMITTED_YET,
     )
 
@@ -1832,7 +1742,11 @@ class WhatsAppTemplate(
         revision.save(update_fields=["content"])
         return revision
 
-    def check_matching_braces(self, message=message):
+    def check_matching_braces(self, message: str = message) -> str:
+        """
+        Check if the number of opening and closing braces match in the message.
+        Returns an error message if they don't match, otherwise returns an empty string.
+        """
         result = ""
         count_opening_braces = message.count("{{")
         count_closing_braces = message.count("}}")
@@ -1842,9 +1756,9 @@ class WhatsAppTemplate(
 
         return result
 
-    def clean(self):
+    def clean(self) -> None:
         result = super().clean()
-        errors = {}
+        errors: dict[str, list[ValidationError]] = {}
 
         # The name is needed for all templates to generate a name for the template
         if not self.name:
@@ -1858,8 +1772,8 @@ class WhatsAppTemplate(
                 errors["example_values"] = ValidationError(
                     "Example values cannot contain commas"
                 )
-
         message = self.message
+
         # Matches "{1}" and "{11}", not "{{1}", "{a}" or "{1 "
         single_braces = re.findall(r"[^{]{(\d*?)}", message)
         # TODO: Replace with PyParsing